f Move util functions to `test::utils`
