+++ src/test/functional_tests.rs
@@ -1,137 +1,10 @@
+use crate::test::utils::*;
 use crate::test::utils::{expect_event, random_config};
 use crate::{Builder, Error, Event, PaymentDirection, PaymentStatus};
 
-use bitcoin::{Address, Amount, OutPoint, Txid};
-use bitcoind::bitcoincore_rpc::RpcApi;
-use electrsd::bitcoind::bitcoincore_rpc::bitcoincore_rpc_json::AddressType;
-use electrsd::{bitcoind, bitcoind::BitcoinD, ElectrsD};
-use electrum_client::ElectrumApi;
+use bitcoin::Amount;
 
-use std::env;
 use std::time::Duration;
-
-fn setup_bitcoind_and_electrsd() -> (BitcoinD, ElectrsD) {
-	let bitcoind_exe =
-		env::var("BITCOIND_EXE").ok().or_else(|| bitcoind::downloaded_exe_path().ok()).expect(
-			"you need to provide an env var BITCOIND_EXE or specify a bitcoind version feature",
-		);
-	let mut bitcoind_conf = bitcoind::Conf::default();
-	bitcoind_conf.network = "regtest";
-	let bitcoind = BitcoinD::with_conf(bitcoind_exe, &bitcoind_conf).unwrap();
-
-	let electrs_exe = env::var("ELECTRS_EXE")
-		.ok()
-		.or_else(electrsd::downloaded_exe_path)
-		.expect("you need to provide env var ELECTRS_EXE or specify an electrsd version feature");
-	let mut electrsd_conf = electrsd::Conf::default();
-	electrsd_conf.http_enabled = true;
-	electrsd_conf.network = "regtest";
-	let electrsd = ElectrsD::with_conf(electrs_exe, &bitcoind, &electrsd_conf).unwrap();
-	(bitcoind, electrsd)
-}
-
-fn generate_blocks_and_wait(bitcoind: &BitcoinD, electrsd: &ElectrsD, num: usize) {
-	let cur_height = bitcoind.client.get_block_count().expect("failed to get current block height");
-	let address = bitcoind
-		.client
-		.get_new_address(Some("test"), Some(AddressType::Legacy))
-		.expect("failed to get new address");
-	// TODO: expect this Result once the WouldBlock issue is resolved upstream.
-	let _block_hashes_res = bitcoind.client.generate_to_address(num as u64, &address);
-	wait_for_block(electrsd, cur_height as usize + num);
-}
-
-fn wait_for_block(electrsd: &ElectrsD, min_height: usize) {
-	let mut header = match electrsd.client.block_headers_subscribe() {
-		Ok(header) => header,
-		Err(_) => {
-			// While subscribing should succeed the first time around, we ran into some cases where
-			// it didn't. Since we can't proceed without subscribing, we try again after a delay
-			// and panic if it still fails.
-			std::thread::sleep(Duration::from_secs(1));
-			electrsd.client.block_headers_subscribe().expect("failed to subscribe to block headers")
-		}
-	};
-	loop {
-		if header.height >= min_height {
-			break;
-		}
-		header = exponential_backoff_poll(|| {
-			electrsd.trigger().expect("failed to trigger electrsd");
-			electrsd.client.ping().expect("failed to ping electrsd");
-			electrsd.client.block_headers_pop().expect("failed to pop block header")
-		});
-	}
-}
-
-fn wait_for_tx(electrsd: &ElectrsD, txid: Txid) {
-	let mut tx_res = electrsd.client.transaction_get(&txid);
-	loop {
-		if tx_res.is_ok() {
-			break;
-		}
-		tx_res = exponential_backoff_poll(|| {
-			electrsd.trigger().unwrap();
-			electrsd.client.ping().unwrap();
-			Some(electrsd.client.transaction_get(&txid))
-		});
-	}
-}
-
-fn wait_for_outpoint_spend(electrsd: &ElectrsD, outpoint: OutPoint) {
-	let tx = electrsd.client.transaction_get(&outpoint.txid).unwrap();
-	let txout_script = tx.output.get(outpoint.vout as usize).unwrap().clone().script_pubkey;
-	let mut is_spent = !electrsd.client.script_get_history(&txout_script).unwrap().is_empty();
-	loop {
-		if is_spent {
-			break;
-		}
-
-		is_spent = exponential_backoff_poll(|| {
-			electrsd.trigger().unwrap();
-			electrsd.client.ping().unwrap();
-			Some(!electrsd.client.script_get_history(&txout_script).unwrap().is_empty())
-		});
-	}
-}
-
-fn exponential_backoff_poll<T, F>(mut poll: F) -> T
-where
-	F: FnMut() -> Option<T>,
-{
-	let mut delay = Duration::from_millis(64);
-	let mut tries = 0;
-	loop {
-		match poll() {
-			Some(data) => break data,
-			None if delay.as_millis() < 512 => {
-				delay = delay.mul_f32(2.0);
-			}
-
-			None => {}
-		}
-		assert!(tries < 10, "Reached max tries.");
-		tries += 1;
-		std::thread::sleep(delay);
-	}
-}
-
-fn premine_and_distribute_funds(
-	bitcoind: &BitcoinD, electrsd: &ElectrsD, addrs: Vec<Address>, amount: Amount,
-) {
-	generate_blocks_and_wait(bitcoind, electrsd, 101);
-
-	for addr in addrs {
-		let txid = bitcoind
-			.client
-			.send_to_address(&addr, amount, None, None, None, None, None, None)
-			.unwrap();
-		wait_for_tx(electrsd, txid);
-	}
-
-	generate_blocks_and_wait(bitcoind, electrsd, 1);
-}
-
 #[test]
 fn channel_full_cycle() {
 	let (bitcoind, electrsd) = setup_bitcoind_and_electrsd();
